fixes for sphinx docs

+++ bin/notebook-mkdocs.sh
@@ -86,16 +86,26 @@ Last update: {{ git_revision_date_localized }}
 
 EOF
 
+# get files
+cd "$base"
+files=docs/*_files
+filedirs=$(echo $files | sed 's/docs\///g')
+cd "$base"/docs/sphinx
+for n in ${filedirs[@]}
+do
+  ln -s ../$n $n
+done
 
-echo "--> generating sphinx files for docs"
-cp  config/index_head.rst docs/index.rst
-awk < mkdocs.yml -F: 'BEGIN{start=0} ($1=="nav"){start=1} ($1=="plugins"){start=0} (start==1 && $1!="nav"){print $NF}' >> docs/index.rst
-cat config/index_tail.rst >> docs/index.rst
-
+cd "$base"
 echo "--> generating mkdocs files for docs"
 geog0111/mkdocs_prep.py --dev
 echo "--> building mkdocs"
-#rm -r docs/index.md
+
+echo "--> generating sphinx files for docs"
+cp  config/index_head.rst docs/index.rst
+awk < mkdocs.yml -F: 'BEGIN{start=0} ($1=="nav"){start=1} ($1=="plugins"){start=0} (start==1 && $1!="nav"){print "  "$NF}' >> docs/index.rst
+cat config/index_tail.rst >> docs/index.rst
+
 mkdocs build -v
 cd $base/docs
 cp ../config/requirements.txt .
@@ -106,11 +116,16 @@ ln -s ../data data
 ln -s ../work work
 ln -s ../geog0111 geog0111
 ln -s ../copy copy
-mv index.md docindex.md
-sed < index.rst 's/index.md/docindex.md/' > tmp.$$
+sed < $base/docs/index.rst 's/index.md/docindex.md/' > tmp.$$
 mv tmp.$$ index.rst
-
+rm -f index.md
 cp ../*.md ../*.html .
+if [ -f "index.md" ]; then
+  mv index.md docindex.md
+else
+  cp ../index.md docindex.md
+fi
+
 #cp ../index.rst .
 make clean html
 cd $base
